+++ SoftwareEngineering/.metadata/.plugins/org.eclipse.m2e.logback/0.log
@@ -1,0 +1,2 @@
+2024-10-14 09:25:44,072 [Worker-5: Loading available Gradle versions] INFO  o.e.b.c.i.u.g.PublishedGradleVersions - Gradle version information cache is out-of-date. Trying to update.
+2024-10-21 10:22:45,018 [Worker-7: Loading available Gradle versions] INFO  o.e.b.c.i.u.g.PublishedGradleVersions - Gradle version information cache is out-of-date. Trying to update.
